Rewrote velocity_verlet_2, has to replace version 1

+++ Thema1_N-Body-Problem/lib/beschleunigung.cpp
@@ -7,23 +7,37 @@ using namespace std;
 
 
 vector<vector<double>> acceleration(vector<vector<double>> table){
+    //initialize vectors and output
     vector<double> a_i;
     vector<double> a_i_one_term;
     vector<vector<double>> acc_matrix;
+
     for(vector<double> particle_i : table){
+        //set a_i to 0
         a_i = {0,0,0};
         a_i_one_term = {0,0,0};
-        //vector<double> r_i = {particle_i[0],particle_i[1],particle_i[2]};
+
+        //set r_i
         vector<double> r_i = extract_position(particle_i);
+
         for(vector<double> particle_j : table){
+
+            //check if i and j are the same (elementwise)
             if(particle_i == particle_j){
                 continue;
             }
-            //vector<double> r_j = {particle_j[0],particle_j[1],particle_j[2]};
+
+            //set r_j, r_ij
             vector<double> r_j = extract_position(particle_j);
             vector<double> r_ij = subtract_vectors(r_j,r_i);
+
+            //calc denominator r_ij^3
             double r_ij_absolute = pow(absolute_value(r_ij), 3);
-            a_i_one_term = scalar_multiplication((particle_j[6]/r_ij_absolute), r_ij);
+
+            //calculate acceleration from particle j
+            a_i_one_term = scalar_multiplication(((double)particle_j[6]/r_ij_absolute), r_ij);
+
+            //add to total accerleration of particle i
             a_i = add_vectors(a_i, a_i_one_term);
              
         }
